fix unlinked files on Windows
add a check for unlink error too. Before the test cases worked on
windows, since the os.Remove error was ignored, but the tempfile was not
deleted. We now check for unlink errors too.

The windows specific code is taken from:

https://github.com/calebcase/tmpfile

we only changed the create flag since the tempfile already exist